# Speed up the function `min_max_string`