# Which issue does this PR close?
Closes #1373 .

# Rationale for this change
 1. Improve performance 
 2. clean up the code

# What changes are included in this PR?
1. Performance is improved when `null_count > 0`. See benchmark result.  
2. more functional styles and fewer mutable variables

## benchmark
```shell
min string 512          time:   [1.3687 us 1.3691 us 1.3696 us]                            
                        change: [-7.7279% -7.5833% -7.4242%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 9 outliers among 100 measurements (9.00%)
  1 (1.00%) high mild
  8 (8.00%) high severe

min nulls string 512    time:   [1.6630 us 1.6647 us 1.6665 us]                                  
                        change: [-33.478% -33.362% -33.244%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) high mild
  1 (1.00%) high severe
```

# Are there any user-facing changes?
No