@HaoYang670 2022-03-01
# Benchmark
## Before
```bash
min string 512          time:   [1.4824 us 1.4914 us 1.5094 us]                            
                        change: [-2.4834% -0.7235% +0.5457%] (p = 0.44 > 0.05)
                        No change in performance detected.
Found 13 outliers among 100 measurements (13.00%)
  2 (2.00%) high mild
  11 (11.00%) high severe

min nulls string 512    time:   [2.4456 us 2.4528 us 2.4610 us]                                  
                        change: [-0.7299% -0.2597% +0.2043%] (p = 0.26 > 0.05)
                        No change in performance detected.
Found 9 outliers among 100 measurements (9.00%)
  7 (7.00%) high mild
  2 (2.00%) high severe
```

## After
```bash
min string 512          time:   [1.5539 us 1.5552 us 1.5565 us]                            
                        change: [-0.4238% -0.2174% -0.0106%] (p = 0.04 < 0.05)
                        Change within noise threshold.
Found 5 outliers among 100 measurements (5.00%)
  1 (1.00%) high mild
  4 (4.00%) high severe

min nulls string 512    time:   [1.8519 us 1.8536 us 1.8552 us]                                  
                        change: [-1.0138% -0.7613% -0.4859%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) low mild
  1 (1.00%) high mild
  1 (1.00%) high severe
```