@HaoYang670 2022-03-01
I have tested that if we replace `array.value` by `array.value_unchecked`, there will be 10% extra performance improvement on the benchmark `min string 512`.  But it also introduces an unsafe block.